+++ CHANGELOG.md
@@ -33,6 +33,7 @@ for further information about branching and tagging conventions.
 - Resolved an issue when disabled or readonly select field was rendered slightly
   differ from other fields
 - Resolved an issue when date field was disabled after submit form
+- Issue with tooltip displaying inside elements with own scroll
 - Resolved an issue when tick field was displayed not chacked if value checked
 
 ### "Under the hood" changes